analyses/anomaly-train-gmm/app/requirements.txt
firebase-admin
google-cloud-pubsub
google-cloud-storage
joblib
python-dateutil
sklearn